requirements.txt
# Core dependencies
openai>=1.0.0
requests>=2.28.0
python-dotenv>=0.19.0

# Image processing
opencv-python>=4.5.0
Pillow>=9.0.0

# Web framework
fastapi>=0.95.0
uvicorn>=0.20.0

# Database
sqlalchemy>=2.0.0
alembic>=1.10.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development
black>=22.0.0
flake8>=5.0.0
pre-commit>=3.0.0

# Monitoring
prometheus-client>=0.16.0
